detection-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pycoral==2.0.0
tflite-runtime==2.14.0
pillow==10.1.0
numpy==1.24.3
opencv-python-headless==4.8.1.78
python-multipart==0.0.6
websockets==12.0
pydantic==2.5.0
python-dotenv==1.0.0
